+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Refactor `ReplicantSessionManagerImpl.saveEntityMessages(...)` so that each session is processed in succession rather than processing all sessions simultaneously an phase-by-phase, event-by-event. This is preparing for a future mechanism via which sessions can be locked during processing to avoid concurrency issues as outlined in stocksoftware/rose#716 and replicant4j/replicant#4.
 * Remove unused `ChangeAccumulator`.
 * Expose `ChangeSet.merge(Collection<Change>)` and `ChangeSet.mergeActions(Collection<Change>)` as public methods.
+* Remove invariant guard that is no longer needed post `v6.71` when invoking `Replicant.context().request(...)`.
 
 ### [v6.73](https://github.com/replicant4j/replicant/tree/v6.73) (2020-02-20) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.72...v6.73)